requirements.txt
tidalapi
plyer
BeautifulSoup4
Pillow
notify-py
